task3_route_optimization/requirements.txt
sentence-transformers==3.1.1
faiss-cpu
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pymongo==4.8.0
numpy
torch
python-multipart==0.0.9
tqdm

